+++ wrappers/NotificationWrapper.jsx
@@ -1,8 +1,0 @@
-"use client";
-import { useEffect, useState } from "react";
-
-import { NotificationsProvider } from "@/context/notificationsContext";
-
-export default function NotificationsWrapper({ children }) {
-  return <NotificationsProvider>{children}</NotificationsProvider>;
-}